feat: SessionRunner — spawn/stop bridges + agents --json cross-check (v0.1 features 2-4)
Manage one `claude remote-control` bridge per project:
- spawn(): detached subprocess + bounded debug-log parse to reach RUNNING
- stop(): mark intent, then SIGINT with procStart+cmdline TOCTOU re-validation
- psutil liveness (zombie/cmdline/create-time guards); bridge-pointer.json
  rediscovery (forward cwd mapping); `claude agents --json` reconcile by cwd
- API: /api/instances CRUD + /api/projects/{name}/trust; dashboard Start/Stop/
  Trust with live status badges
- atomic ~/.claude.json trust writer
Move logfire to an [eval] extra (evaluation only, not core). 64 tests pass.

+++ src/clauster/models.py
@@ -42,8 +42,8 @@ class RemoteControlInstance(BaseModel):
 
     project: str
     label: str  # passed as --name to claude remote-control
-    bridge_pid: int | None = None  # parent PID from subprocess
-    bridge_proc_start: int | None = None  # psutil create-time; PID-reuse defense
+    bridge_pid: int | None = None  # bridge parent PID (matches bridge-pointer pid)
+    bridge_proc_start: float | None = None  # psutil create-time (epoch); PID-reuse defense
     bridge_id: str | None = None  # UUID from bridge:init log
     environment_id: str | None = None  # env_<ULID>; the URL parameter
     starter_session_id: str | None = None  # session_<ULID> from "Created initial session"
